add erupt-print config API

+++ erupt-ai/src/main/java/xyz/erupt/ai/core/LlmRequest.java
@@ -18,9 +18,9 @@ public class LlmRequest {
 
     private String model;
 
-    private Float temperature;
+    private Double temperature;
 
-    private Float top_p;
+    private Double top_p;
 
     private ResponseFormat responseFormat = ResponseFormat.text;
